Remove dead code, apply model dtype to all models, reduce torch conversions

+++ tripy/examples/segment-anything-model-v2/sam2/modeling/backbones/image_encoder.py
@@ -55,7 +55,7 @@ class ImageEncoder(tp.Module):
         else:
             features_pos = self.forward(sample)
         for i in range(len(features_pos)):
-            features_pos[i] = torch.from_dlpack(features_pos[i]).to(torch.float32)
+            features_pos[i] = torch.from_dlpack(features_pos[i])
         n = len(self.neck.backbone_channel_list)
         features = list(features_pos[:n])
         pos = list(features_pos[n:])
